requirements.txt
Flask==3.0.0
pymongo==4.6.1
pandas>=2.2.0
openpyxl==3.1.2
reportlab==4.0.7
gunicorn==21.2.0
python-dotenv==1.0.0
certifi
 